+++ stream-chat-android-ui-components-sample/src/main/kotlin/io/getstream/chat/ui/sample/feature/chat/info/group/GroupChatInfoViewModel.kt
@@ -4,23 +4,35 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import io.getstream.chat.android.client.ChatClient
 import io.getstream.chat.android.client.call.await
 import io.getstream.chat.android.client.channel.ChannelClient
 import io.getstream.chat.android.client.models.ChannelMute
 import io.getstream.chat.android.client.models.Member
-import io.getstream.chat.android.livedata.ChatDomain
+import io.getstream.chat.android.core.internal.InternalStreamChatApi
 import io.getstream.chat.android.livedata.utils.Event
+import io.getstream.chat.android.offline.experimental.channel.state.ChannelState
+import io.getstream.chat.android.offline.experimental.extensions.asReferenced
+import io.getstream.chat.android.offline.experimental.extensions.globalState
+import io.getstream.chat.android.offline.experimental.global.GlobalState
 import io.getstream.chat.android.ui.common.extensions.isOwnerOrAdmin
-import io.getstream.chat.ui.sample.common.name
 import kotlinx.coroutines.launch
 
+@OptIn(InternalStreamChatApi::class)
 class GroupChatInfoViewModel(
     private val cid: String,
-    private val chatDomain: ChatDomain = ChatDomain.instance(),
     private val chatClient: ChatClient = ChatClient.instance(),
+    private val globalState: GlobalState = chatClient.globalState,
 ) : ViewModel() {
+
+    /**
+     * Holds information about the current channel and is actively updated.
+     */
+    @OptIn(InternalStreamChatApi::class)
+    private val channelState: ChannelState =
+        chatClient.asReferenced().watchChannel(cid).asState(viewModelScope)
 
     private val channelClient: ChannelClient = chatClient.channel(cid)
     private val _state = MediatorLiveData<State>()
@@ -32,25 +44,21 @@ class GroupChatInfoViewModel(
 
     init {
         _state.value = INITIAL_STATE
-        chatDomain.getChannelController(cid).enqueue { result ->
-            if (result.isSuccess) {
-                val controller = result.data()
-                // Update channel mute status
-                chatDomain.user.value?.channelMutes?.let(::updateChannelMuteStatus)
 
-                // Update members
-                _state.addSource(controller.members, this::updateMembers)
+        // Update channel mute status
+        globalState.user.value?.channelMutes?.let(::updateChannelMuteStatus)
 
-                getOwnerOrAdmin(controller.members.value)?.let { member ->
-                    _state.value = _state.value?.copy(
-                        isCurrentUserOwnerOrAdmin = chatDomain.user.value?.id == member.getUserId()
-                    )
-                }
+        // Update members
+        _state.addSource(channelState.members.asLiveData(), this::updateMembers)
 
-                _state.addSource(controller.offlineChannelData) { channelData ->
-                    _state.value = _state.value?.copy(channelName = channelData.name)
-                }
-            }
+        getOwnerOrAdmin(channelState.members.value)?.let { member ->
+            _state.value = _state.value?.copy(
+                isCurrentUserOwnerOrAdmin = globalState.user.value?.id == member.getUserId()
+            )
+        }
+
+        _state.addSource(channelState.channelData.asLiveData()) { channelData ->
+            _state.value = _state.value?.copy(channelName = channelData.name)
         }
     }
 
@@ -72,7 +80,7 @@ class GroupChatInfoViewModel(
     }
 
     private fun handleMemberClick(member: Member) {
-        if (member.getUserId() != chatDomain.user.value?.id) {
+        if (member.getUserId() != globalState.user.value?.id) {
             val currentState = _state.value!!
             _events.value = Event(UiEvent.ShowMemberOptions(member, currentState.channelName))
         }