+++ CHANGELOG.md
@@ -50,9 +50,12 @@
 ### ✅ Added
 
 ### ⚠️ Changed
+- Removed `ChatDomain` from all `ViewModels`. Now they rely on `OfflinePlugin` for offline capabilities. [#3190](https://github.com/GetStream/stream-chat-android/pull/3190)
+- Removed all uses of `ChatDomain` from all `Views` and functions that relied on it. Now they rely on `OfflinePlugin` for offline capabilities. [#3190](https://github.com/GetStream/stream-chat-android/pull/3190) 
 
 ### ❌ Removed
 - Removed ChatMarkdown in favor of ChatMessageTextTransformer [#3189](https://github.com/GetStream/stream-chat-android/pull/3189)
+- All usage of `ChatDomain`. [#3190](https://github.com/GetStream/stream-chat-android/pull/3190)
 
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
@@ -76,9 +79,12 @@
 - Renamed MediaAttachmentViewStyle and its attributes to ImageAttachmentViewStyle. [#3067](https://github.com/GetStream/stream-chat-android/pull/3067)
 - Messages containing link attachments and no other types of attachments are no longer handled by `TextAndAttachmentsViewHolder`, instead they are handled by `LinkAttachmentsViewHolder`. [#3070](https://github.com/GetStream/stream-chat-android/pull/3070)
 - Messages containing file attachments of different file types or types not handled by one of the other `ViewHolders` are no longer handled by `TextAndAttachmentsViewHolder`, instead they are handled by `FileAttachmentsViewHolder`. [#3091](https://github.com/GetStream/stream-chat-android/pull/3091)
+- Removed `ChatDomain` from all `ViewModels`. Now they rely on `OfflinePlugin` for offline capabilities. [#3190](https://github.com/GetStream/stream-chat-android/pull/3190)
+- Removed all uses of `ChatDomain` from all `Views` and functions that relied on it. Now they rely on `OfflinePlugin` for offline capabilities. [#3190](https://github.com/GetStream/stream-chat-android/pull/3190)
 - Updated the structure of UI components documentation. [UI Components documentation](https://getstream.io/chat/docs/sdk/android/ui/overview/). [#3186](https://github.com/GetStream/stream-chat-android/pull/3186)
 
 ### ❌ Removed
+- All usage of `ChatDomain`. [#3190](https://github.com/GetStream/stream-chat-android/pull/3190)
 
 ## stream-chat-android-compose
 ### 🐞 Fixed
@@ -88,6 +94,7 @@
 - Allowed passing long press handlers to `MessageText`. [#3137](https://github.com/GetStream/stream-chat-android/pull/3137)
 
 ### ✅ Added
+- Added code snippets from the Compose documentation to the `stream-chat-android-docs` module. [3197](https://github.com/GetStream/stream-chat-android/pull/3197)
 
 ### ⚠️ Changed
 - Replaced the `reactionTypes` field in `ChatTheme` with the new `reactionIconFactory` field that allows customizing reaction icons. [#3046](https://github.com/GetStream/stream-chat-android/pull/3046)